fix: move kitchen-sink event stream tests into parent modules
Drop per-client internal/testing submodules that required their no_tag
parent, which tripped calculaterelease. Tests now live in parent modules
as white-box tests. Regenerated on schema-serde cbor wave 2.
