+++ internal/kitchensinktest/eventstream_test.go
@@ -1,4 +1,4 @@
-package testing
+package kitchensinktest
 
 import (
 	"bytes"
@@ -11,7 +11,6 @@ import (
 
 	"github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream"
 	"github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream/eventstreamapi"
-	kitchensink "github.com/aws/aws-sdk-go-v2/internal/kitchensinktestrestjsonlegacy"
 	smithyendpoints "github.com/aws/smithy-go/endpoints"
 	"github.com/aws/smithy-go/middleware"
 	smithyhttp "github.com/aws/smithy-go/transport/http"
@@ -34,7 +33,7 @@ func (b *closeTrackingBody) closeCount() int { return int(atomic.LoadInt32(&b.cl
 
 type eventStreamEndpointResolver struct{}
 
-func (*eventStreamEndpointResolver) ResolveEndpoint(ctx context.Context, params kitchensink.EndpointParameters) (smithyendpoints.Endpoint, error) {
+func (*eventStreamEndpointResolver) ResolveEndpoint(ctx context.Context, params EndpointParameters) (smithyendpoints.Endpoint, error) {
 	return smithyendpoints.Endpoint{URI: url.URL{Scheme: "https", Host: "test.example.com"}}, nil
 }
 
@@ -52,14 +51,12 @@ func encodeEventStream(t *testing.T, msgs []eventstream.Message) []byte {
 	return buf.Bytes()
 }
 
-func strPtr(s string) *string { return &s }
-
 // TestSubscribeEvents_Close verifies the caller-owned event stream behavior on
-// the success path: the response body is handed to the event stream reader
-// rather than closed by the deserialize path, events are readable, and closing
-// the stream via GetStream().Close() closes the underlying body exactly once.
-// This locks in the pre-existing behavior so a middleware-stack change can be
-// shown not to alter it.
+// the success path for the awsJson protocol: the response body is handed to the
+// event stream reader rather than closed by the deserialize path, events are
+// readable, and closing the stream via GetStream().Close() succeeds. awsJson
+// (HttpRpc codegen path) is independent from the restJson1 (HttpBinding path)
+// implementation, so it is exercised separately.
 func TestSubscribeEvents_Close(t *testing.T) {
 	payload := encodeEventStream(t, []eventstream.Message{
 		{
@@ -80,7 +77,7 @@ func TestSubscribeEvents_Close(t *testing.T) {
 
 	body := &closeTrackingBody{r: bytes.NewReader(payload)}
 
-	svc := kitchensink.New(kitchensink.Options{
+	svc := New(Options{
 		Region: "us-east-1",
 		HTTPClient: smithyhttp.ClientDoFunc(func(req *http.Request) (*http.Response, error) {
 			h := http.Header{}
@@ -102,9 +99,7 @@ func TestSubscribeEvents_Close(t *testing.T) {
 		},
 	})
 
-	resp, err := svc.SubscribeEvents(context.Background(), &kitchensink.SubscribeEventsInput{
-		Id: strPtr("some-id"),
-	})
+	resp, err := svc.SubscribeEvents(context.Background(), &SubscribeEventsInput{})
 	if err != nil {
 		t.Fatalf("subscribe events: %v", err)
 	}
