--- internal/kitchensinktestlegacy/internal/testing/go.mod
@@ -1,45 +1,0 @@
-module github.com/aws/aws-sdk-go-v2/internal/kitchensinktestlegacy/internal/testing
-
-go 1.24
-
-require (
-	github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream v1.7.14
-	github.com/aws/aws-sdk-go-v2/internal/kitchensinktestlegacy v0.0.0-00010101000000-000000000000
-	github.com/aws/smithy-go v1.27.3
-)
-
-require (
-	github.com/aws/aws-sdk-go-v2 v1.43.0 // indirect
-	github.com/aws/aws-sdk-go-v2/internal/configsources v1.4.31 // indirect
-	github.com/aws/aws-sdk-go-v2/internal/endpoints/v2 v2.7.31 // indirect
-)
-
-replace github.com/aws/aws-sdk-go-v2 => ../../../../
-
-replace github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream => ../../../../aws/protocol/eventstream/
-
-replace github.com/aws/aws-sdk-go-v2/credentials => ../../../../credentials/
-
-replace github.com/aws/aws-sdk-go-v2/feature/ec2/imds => ../../../../feature/ec2/imds/
-
-replace github.com/aws/aws-sdk-go-v2/internal/configsources => ../../../../internal/configsources/
-
-replace github.com/aws/aws-sdk-go-v2/internal/endpoints/v2 => ../../../../internal/endpoints/v2/
-
-replace github.com/aws/aws-sdk-go-v2/internal/kitchensinktestlegacy => ../../
-
-replace github.com/aws/aws-sdk-go-v2/internal/v4a => ../../../../internal/v4a/
-
-replace github.com/aws/aws-sdk-go-v2/service/internal/accept-encoding => ../../../../service/internal/accept-encoding/
-
-replace github.com/aws/aws-sdk-go-v2/service/internal/eventstreamtesting => ../../../../service/internal/eventstreamtesting/
-
-replace github.com/aws/aws-sdk-go-v2/service/internal/presigned-url => ../../../../service/internal/presigned-url/
-
-replace github.com/aws/aws-sdk-go-v2/service/signin => ../../../../service/signin/
-
-replace github.com/aws/aws-sdk-go-v2/service/sso => ../../../../service/sso/
-
-replace github.com/aws/aws-sdk-go-v2/service/ssooidc => ../../../../service/ssooidc/
-
-replace github.com/aws/aws-sdk-go-v2/service/sts => ../../../../service/sts/
