+++ internal/kitchensinktestrestjson/eventstream_test.go
@@ -1,40 +1,22 @@
-package testing
+package kitchensinktestrestjson
 
 import (
 	"bytes"
 	"context"
-	"io"
 	"net/http"
 	"net/url"
-	"sync/atomic"
 	"testing"
 
 	"github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream"
 	"github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream/eventstreamapi"
-	kitchensink "github.com/aws/aws-sdk-go-v2/internal/kitchensinktestlegacy"
 	smithyendpoints "github.com/aws/smithy-go/endpoints"
 	"github.com/aws/smithy-go/middleware"
 	smithyhttp "github.com/aws/smithy-go/transport/http"
 )
 
-// closeTrackingBody wraps a response body and records whether Close was called.
-type closeTrackingBody struct {
-	r      io.Reader
-	closed int32
-}
-
-func (b *closeTrackingBody) Read(p []byte) (int, error) { return b.r.Read(p) }
-
-func (b *closeTrackingBody) Close() error {
-	atomic.AddInt32(&b.closed, 1)
-	return nil
-}
-
-func (b *closeTrackingBody) closeCount() int { return int(atomic.LoadInt32(&b.closed)) }
-
 type eventStreamEndpointResolver struct{}
 
-func (*eventStreamEndpointResolver) ResolveEndpoint(ctx context.Context, params kitchensink.EndpointParameters) (smithyendpoints.Endpoint, error) {
+func (*eventStreamEndpointResolver) ResolveEndpoint(ctx context.Context, params EndpointParameters) (smithyendpoints.Endpoint, error) {
 	return smithyendpoints.Endpoint{URI: url.URL{Scheme: "https", Host: "test.example.com"}}, nil
 }
 
@@ -53,11 +35,11 @@ func encodeEventStream(t *testing.T, msgs []eventstream.Message) []byte {
 }
 
 // TestSubscribeEvents_Close verifies the caller-owned event stream behavior on
-// the success path for the awsJson protocol: the response body is handed to the
-// event stream reader rather than closed by the deserialize path, events are
-// readable, and closing the stream via GetStream().Close() succeeds. awsJson
-// (HttpRpc codegen path) is independent from the restJson1 (HttpBinding path)
-// implementation, so it is exercised separately.
+// the success path: the response body is handed to the event stream reader
+// rather than closed by the deserialize path, events are readable, and closing
+// the stream via GetStream().Close() closes the underlying body exactly once.
+// This locks in the pre-existing behavior so a middleware-stack change can be
+// shown not to alter it.
 func TestSubscribeEvents_Close(t *testing.T) {
 	payload := encodeEventStream(t, []eventstream.Message{
 		{
@@ -78,7 +60,7 @@ func TestSubscribeEvents_Close(t *testing.T) {
 
 	body := &closeTrackingBody{r: bytes.NewReader(payload)}
 
-	svc := kitchensink.New(kitchensink.Options{
+	svc := New(Options{
 		Region: "us-east-1",
 		HTTPClient: smithyhttp.ClientDoFunc(func(req *http.Request) (*http.Response, error) {
 			h := http.Header{}
@@ -100,7 +82,9 @@ func TestSubscribeEvents_Close(t *testing.T) {
 		},
 	})
 
-	resp, err := svc.SubscribeEvents(context.Background(), &kitchensink.SubscribeEventsInput{})
+	resp, err := svc.SubscribeEvents(context.Background(), &SubscribeEventsInput{
+		Id: strPtr("some-id"),
+	})
 	if err != nil {
 		t.Fatalf("subscribe events: %v", err)
 	}
