+++ internal/kitchensinktestrestjson/go.mod
@@ -4,6 +4,7 @@ go 1.24
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.43.0
+	github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream v1.7.14
 	github.com/aws/aws-sdk-go-v2/internal/configsources v1.4.31
 	github.com/aws/aws-sdk-go-v2/internal/endpoints/v2 v2.7.31
 	github.com/aws/smithy-go v1.27.3
