+++ modman.toml
@@ -76,25 +76,13 @@
   [modules."internal/kitchensinktest"]
     no_tag = true
 
-  [modules."internal/kitchensinktest/internal/testing"]
-    no_tag = true
-
   [modules."internal/kitchensinktestlegacy"]
-    no_tag = true
-
-  [modules."internal/kitchensinktestlegacy/internal/testing"]
     no_tag = true
 
   [modules."internal/kitchensinktestrestjson"]
     no_tag = true
 
-  [modules."internal/kitchensinktestrestjson/internal/testing"]
-    no_tag = true
-
   [modules."internal/kitchensinktestrestjsonlegacy"]
-    no_tag = true
-
-  [modules."internal/kitchensinktestrestjsonlegacy/internal/testing"]
     no_tag = true
 
   [modules."internal/repotools"]
